requirements.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.1
Django==3.1.4
django-appconf==1.0.4
django-cors-headers==3.6.0
django-phonenumber-field==5.0.0
django-redis==4.12.1
django-select2==7.5.0
djangorestframework==3.12.2
filelock==3.0.12
isort==5.6.4
phonenumbers==8.12.15
Pillow==8.0.1
python-dotenv==0.15.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
virtualenv==20.2.2
